fix shell script

+++ build.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-rm -rf build && parcel build index.html --out-dir build
+if [ -d "./build" ]
+then
+    rm -rf build && parcel build index.html --out-dir build
+else
+    parcel build index.html --out-dir build
+fi
